devel/py-isoduration: Add py-isoduration 20.11.0
ISO 8601 is most commonly known as a way to exchange datetimes in textual
format. A lesser known aspect of the standard is the representation of
durations. They have a shape similar to this:
    P3Y6M4DT12H30M5S
This string represents a duration of 3 years, 6 months, 4 days, 12 hours, 30
minutes, and 5 seconds.

The state of the art of ISO 8601 duration handling in Python is more or less
limited to what's offered by isodate. What we are trying to achieve here is to
address the shortcomings of isodate (as described in their own Limitations
section), and a few of our own annoyances with their interface, such as the lack
of uniformity in their handling of types, and the use of regular expressions for
parsing.