Mk/Uses/python.mk: Add support for post-release version (.postX)
The post-release version is normalized to .postX in PEP440. However, it will be
converted to .pX in FreeBSD which means an older version.

% pkg ver -t 1.2.3 1.2.3.p4
>

If the original release is already in the tree, rather than bumping PORTEPOCH,
you could bump PORTREVISION and add .postX to DISTVERSIONSUFFIX.

This fix allows the port to build in this situation.