www/nginx-devel: is able to build with OpenSSL compatibility layer now
Here's the quote from README file.

A library that provides QUIC support is recommended to build nginx,
there are several of those available on the market:
 o) BoringSSL
 o) LibreSSL
 o) QuicTLS

Alternatively, nginx can be configured with OpenSSL compatibility
layer, which emulates BoringSSL QUIC API for OpenSSL.  This mode is
enabled by default if native QUIC support is not detected.
0-RTT is not supported in OpenSSL compatibility mode.

Bump PORTREVISION.