Mk/bsd.port.mk: patch rawmemchr in embedded Gnulib
The longword data type in rawmemchr() from Gnulib aims to store as many
bytes as a pointer can. The definition of this data type was changed
from unsigned long int to uintptr_t in [0]. However, on CHERI
architectures, uintptr_t does not store the expected number of bytes.

Revert the longword data type definition to unsigned long int to extend
support to CHERI architectures.

[0] http://git.savannah.gnu.org/gitweb/?p=gnulib.git;a=commitdiff;h=b8ae80b252a4638c52c8f03bbf16b9c8d6bf27de