net-mgmt/fence-agents: new port
Fence agents were developed as device "drivers" which are able to prevent
computers from destroying data on shared storage.
Their aim is to isolate a corrupted computer, using one of three methods:

- Power - A computer that is switched off cannot corrupt data, but it is
  important to not do a "soft-reboot" as we won't know if this is possible.
  This also works for virtual machines when the fence device is a hypervisor.
- Network - Switches can prevent routing to a given computer, so even if a
  computer is powered on it won't be able to harm the data.
- Configuration - Fibre-channel switches or SCSI devices allow us to limit who
  can write to managed disks.

This is used by the Pacemaker cluster resource management.

Note:
A lot of these agents have Linuxism and have to be adapted, they will be
patched later to work out of the box on FreeBSD.

PR:		270623
Sponsored by:	OVHcloud