graphics/gdk-pixbuf2-xlib: remove useless trigger
The trigger provided by gdk-pixbuf-xlib is the same as the one
provided by gdk-pixbuf2 itself, gtk-pixbuf-xlib actually depends on
gdk-pixbuf2 so it will anyway trigger the trigger.

Having 2 triggers do make the same commandes being executed twice
when only once is enough

The only difference between the 2 is the fact the cleanup version
in xlib one to prevent stepping on the toes of the gdk-pixbuf2.