devel/py-requirements-detector: Add py-requirements-detector 1.1.0
Requirements Detector is a simple Python tool which attempts to find and list
the requirements of a Python project.

When run from the root of a Python project, it will try to ascertain which
libraries and the versions of those libraries that the project depends on.

It uses the following methods in order, in the root of the project:
 1. Parse setup.py (if this is successful, the remaining steps are skipped)
 2. Parse pyproject.yoml (if a tool.poetry.dependencies section is found, the
    remaining steps are skipped)
 3. Parse requirements.txt or requirements.pip
 4. Parse all *.txt and *.pip files inside a folder called requirements
 5. Parse all files in the root folder matching *requirements*.txt or reqs.txt
    (so for example, pip_requirements.txt would match, as would
    requirements_common.txt)