sysutils/cpu-x: flavorize the port (by popular demand)
CPU-X can be used in graphic mode with GTK+3, or in text-based mode
with ncurses, subject to build-time configuration.  Previously this
was controlled via options, but since graphic mode was the default,
users who preferred ncurses (console) version had to reconfigure the
port and build their own package, which was inconvenient.  Now with
flavors, everyone should be happy.

Requested by:	marck et al.