lang/ghc: Starting with GHC 9.2.7 switch to the Hadrian build system.
Hadrian is a new build system for GHC which is written in Haskell and requires
bootstrapping process. Ugh. Luckily, we only prepare a source bootstrap
archive as bootstrap GHC has everything needed to build it.

Hadrian also fails to build GHC on i386 so build it the old way for now.