net/wifi-firmware-*: add framework and firmware for various WiFi chipsets
Add a "base port framework" to fetch (WiFi) firmware and license files
from the linux-firmware.git repository and an accompanying FreeBSD
build kmod framework tarball (for now; we might do without that in
the future).
Add child ports for Realtek rtw88 and rtw89, Mediatek mt76, and
Atheros ath10k and ath11k firmware as a start.

This is a beginning in order to stop bloating src.git and the
installed kernels with lots of MB of firmware files whereas usually
only one vebdor/chipset/flavor is needed.  Before fully switching
more infrastructure needs to be in place but this is the first and
most important step.

PR:		270742
Reviewed by:	lwhsu
Approved by:	lwhsu (ports)
Differential Revision: https://reviews.freebsd.org/D39487