devel/py-datasette: Add py-datasette 0.64.2
Datasette is a tool for exploring and publishing data. It helps people take data
of any shape or size and publish that as an interactive, explorable website and
accompanying API.

Datasette is aimed at data journalists, museum curators, archivists, local
governments, scientists, researchers and anyone else who has data that they wish
to share with the world.