www/py-dj41-django-redis: Change RUN_DEPENDS back to py-redis
The previous commit 957fc72127d3 changed the RUN_DEPENDS from py-redis
to py-redis3, which was derived from the set version constraint (= <4).

The constraint was added as a precautionary measure when this port was
introduced to the ports tree due an currently open issue in conjunction
with py-redis >= 4 (issue #595) and therefore to avoid possible
problems/regressions in regard to net-mgmt/netbox.

The last change to this port has now led to installation conflicts with
net-mgmt/netbox because it requires this port as well as
devel/py-dj41-django-rq.  The latter one, pulls in databases/py-redis,
which in the end led to the reported problem. [1]

Fix the issue by switching back to py-redis and also removing the
version constraint.

PR:		270792
Reported by:	Fredrik Eriksson [1]
Approved by:	sunpoet (maintainer)