graphics/opencolorio: Fix BUILD
After textproc/py-sphinx_press_theme was updated it merged some of it's
html and css stuffs in other files and removes two other files that were
generated in the build of opencolorio. This patch fixes the build. Also
bump so that the package is newly built with the latest
textproc/py-sphinx_press_theme.

Reported by:	antoine