sysutils/slurm-wlm: Unbreak build: add perl5 to the build environment
It's not clear to me how this ever built given all the places in
configure that use the "perl" command. But it was building (for me
at least) as recently as yesterday. I can only assume something
changed in the ports build infrastructure.

Anyway, this lets me build kicad again.