sysutils/dynamic_motd: -- new port
dynamic_motd is a drop-in replacement for generating messages of the day on
FreeBSD 13.0 or later. During system startup, the file /var/run/motd is
replaced with a fifo whose content is generated by a user-configurable shell
script that is executed on each login. All other aspects of motd(5) are
preserved.

PR:		267107
Reported by:	Steven Stallion