devel/py-versioneer: Add py-versioneer 0.28
Versioneer is a tool for managing a recorded version number in setuptools-based
python projects. The goal is to remove the tedious and error-prone "update the
embedded version string" step from your release process. Making a new release
should be as easy as recording a new tag in your version-control system, and
maybe making new tarballs.