devel/py-donut-shellcode: New port: donut shellcode Python C extension
Donut is a position-independent code that enables in-memory execution of
VBScript, JScript, EXE, DLL files and dotNET assemblies. A module created by
Donut can either be staged from a HTTP server or embedded directly in the
loader itself. The module is optionally encrypted using the Chaskey block
cipher and a 128-bit randomly generated key. After the file is loaded and
executed in memory, the original reference is erased to deter memory scanners.