math/octave-forge-pkg-octave-doc: New port.
This package provides functions for generating HTML pages that
contain the help texts of the functions of an octave package. The
package is designed to work with installed packages and use their
INDEX file for creating the respective functions' HTML pages. The
default layout is based on boootstrap 5 and it follows the design of
the Octave Packages GitHub page.