devel/qt5-qmake: do not try to install qtconfig header
Qt-dist ports that are a module need to install their header
to qt-config.h. This accidentally happens to qt5-qmake as well.

This is only visible when *installing* qt5-qmake on an otherwise
fresh system with no other Qt packages. If you install any other
Qt packages, then the qt-config header is there for the
post-install script to work on (though it doesn't do anything
for qt5-qmake). In a clean system, the script fails.

PR:		269232
Reported by:	Zsolt Udvari