*/*: Move consumers of libebur128 to ebur128
ebur128 provides a noticable faster implementation on at least aarch64
and amd64 so make it default treewide as they conflict. If there's a
need further down the road we can add an option to choose between.

PR:		271175
Approved by:	acm, jhale (maintainers of some of the ports)
Sponsored by:	Blinkinblox