net-mgmt/nrpe: Resurrect port with new version
THIS WILL REPLACE net-mgmt/nrpe3 WHICH FAILS TO BUILD WITH OpenSSL 3.0.0
and later.

nrpe is used to execute Nagios plugins on remote hosts and report the
results to the main Nagios server. From the Nagios homepage:

Allows you to execute "local" plugins (like check_disk, check_procs,
etc.) on remote hosts. The check_nrpe plugin is called from Nagios and
actually makes the plugin requests to the remote host. Requires that
nrpe be running on the remote host (either as a standalone daemon or as
a service under inetd).