net/mpd5: undo regression for specific PPPoE peers
mpd5-5.9_14 introduced regression for specific case
when no mtu configured for PPPoE peer that does not advertize any MTU itself.
mpd5 should default to 1492 then but revision 14 defaults to 1500.

The problem may be worked around specifying 1492 for link or interface MTU
until upgrade to this revision that restores compatibility.

PR:	270687