net/wifi-firmware-*: use metaport for "base" and switch to Makefile.inc
Rename the wifi-firmware-base to wifi-firmware-kmod (similar to
gpu-firmware-mod).  The former was never build so technically it
is no move.  Make the new subdir a metaport including all
wireless firmware.
Move the old Makefile to Makefile.inc and adjust the child ports
accordingly.

This solves two problems: (1) the base port is now buildable and (2)
there is a single port to depend on to add all the needed firmware
files to, e.g., put on a release media.

Suggested by:	amdmi3
PR:		270742
Reviewed by:	ehaupt, jrm
Approved by:	ehaupt, jrm
Differential Revision: https://reviews.freebsd.org/D40014