sysutils/py-ansible-lint: update to 6.16.1
Loosen version requirements for the RUN_DEPENDS on py-ruamel.yaml, but
tighten them for the TEST_DEPENDS on that package.  This seems
questionable to me, but it is following upstream, and in the
particular case of the ports, the py-ruamel.yaml package now fulfils
both sets of version requirements.

Changes:	https://github.com/ansible/ansible-lint/releases/tag/v6.16.1
Reported by:	repology