graphics/ikona: mark broken and deprecated, abandoned
This hasn't compiled successfully in over a month,
since it comes with Rust code that is no longer compatible
with the current language. There's a couple of commits
after the last upstream tag, but the project is abandoned
upstream. kde@ doesn't have resources to even try to patch this.

Mark for removal before the next quarterly.