mail/mailpit: Update to 1.6.10
Also really make it compile on all arches:

Mailpit uses esbuild via npm to compile frontend assets. Esbuild
is distributed in binary form. There is a webassembly version and
I replaced esbuild with the webassembly version. The port was still
failing because npm was trying to install binary esbuild anyway as
a peer dependency of another requirement. DIsabling peer dependencies
with the "--legacy-peer-deps" allows the build to succeed.