sysutils/ncdu2: Add ncdu2 2.2.2
Ncdu is a disk usage analyzer with an ncurses interface. It is designed to find
space hogs on a remote server where you don't have an entire graphical setup
available, but it is a useful tool even on regular desktop systems. Ncdu aims to
be fast, simple and easy to use, and should be able to run in any minimal
POSIX-like environment with ncurses installed.