net/py-trio-websocket: Add py-trio-websocket 0.10.2
Trio WebSocket implements both server and client aspects of the the WebSocket
protocol, striving for safety, correctness, and ergonomics. It is based on the
wsproto project, which is a Sans-IO state machine that implements the majority
of the WebSocket protocol, including framing, codecs, and events. This library
handles I/O using the Trio framework. This library passes the Autobahn Test
Suite.