ports-mgmt/portmaster: update to version 3.25
The previous update attempted to fix the deletion of stale distfiles
that have names with blank characters or shell glob characters.
It introduced a regression that affected the deletion of some files,
and it has been found to be incomplete.

Testing revealed that the stale distfile check could start before the
full list of relevant distfiles had been read in the background. This
could have lead to distfiles being removed that are still relevant for
other ports.

This update should contain fixes for all these issues.

Reported by:	Dan Allen <danallen46@gmail.com>