ports-mgmt/portmaster: Make --clean-packages consider port flavors
This commit fixes GutHub issue #55 which affected the cleaning of
stale package files for  flavored ports.

Some ports have package names that are prefixed by e.g. an interpreter
version (or where the interpreter version follows the name), but do
not support flavors. Such packages are not dealt with by this update,
since there is no generic way to do so.

The fix does not apply to --clean-packages used with --index or
--index-only, since the	INDEX file does	not record port flavors.