graphics/tippecanoe: new port
Tippecanoe Builds vector tilesets from large (or small) collections of
GeoJSON, FlatGeobuf, or CSV features.

The goal of Tippecanoe is to enable making a scale-independent view of
your data, so that at any level from the entire world to a single
building, you can see the density and texture of the data rather than a
simplification from dropping supposedly unimportant features or
clustering or aggregating them.

If you give it all of OpenStreetMap and zoom out, it should give you
back something that looks like "All Streets" rather than something that
looks like an Interstate road atlas.

If you give it all the building footprints in Los Angeles and zoom out
far enough that most individual buildings are no longer discernable, you
should still be able to see the extent and variety of development in
every neighborhood, not just the largest downtown buildings.

If you give it a collection of years of tweet locations, you should be
able to see the shape and relative popularity of every point of interest
and every significant travel corridor.

WWW: https://github.com/felt/tippecanoe

PR:		266761