Mk/Scripts: Fix the 'stripped' check from 'make check-plist' to report all unstripped files
Prior to this patch, the 'stripped' check always skipped the first
unstripped file.

It uses the "find [...] -exec sh -c 'readelf -S -- /dev/null $0 "$@" || :' -- {} +"
command. When arguments are passed to shell like this:
"sh -c 'script' arg1 arg2 arg3" - $@ within the script is assigned
to 'arg2 arg3', and $0 is assigned to arg1. This is a quirk in
how shells handle arguments in case when the script is passed
using -c.

This patch adds $0 to account for the first passed file.

PR:		270398
Approved by:	portmgr-lurkers (garga)