math/Imath: update → 3.1.9, fixing library versioning
Note this reverts the .so file major version from 31 to 29,
but after some discussion, the AcademySoftwareFoundation people
seem to have revised the versioning policies and also did not
intentionally make breaking changes.
<https://github.com/AcademySoftwareFoundation/Imath/issues/189>

It looks as though future releases should be less painful
WRT versioning.

Unfortunately that means we need to bump PORTREVISION and rebuild ports
again, and it's getting messy with respect to 2023Q2, which has
3.1.7 and major version 30...

Changelog:
https://github.com/AcademySoftwareFoundation/Imath/blob/v3.1.9/CHANGES.md#version-319-May-31-2023