graphics/qt5-imageformats: fix build with clang 16
Since clang 16 (and gcc 11) the default C++ standard is now gnu++17.
Because graphics/qt5-imageformats 's build infrastructure does not
explicitly set its C++ standard, this leads to an issue detecting libmng
in the configure stage:

  Configure summary:

  Further Image Formats:
    JasPer ................................. yes
    MNG .................................... no
    TIFF ................................... yes
      Using system libtiff ................. yes
    WEBP ................................... yes
      Using system libwebp ................. yes

This is because the small configure test program for libmng includes
libmng.h, which transitively includes lcms.h, and the latter fails due
to 'register' keywords in the header. With lcms2.h there is a special
define CMS_NO_REGISTER_KEYWORD to avoid usage of the register keyword,
but lcms.h is very old and does not support that.

As a workaround, define the register keyword away via a "head" object in
src/imageformats/configure.json, and via CXXFLAGS for the rest of the
build.

PR:		271914
Approved by:	tcberner (maintainer)
MFH:		2023Q2